complete multiple arguments

+++ commands/util/clean.js
@@ -18,11 +18,15 @@ module.exports = {
         return message.channel.send(message.author + ' Please use numbers!');
       }
     } else {
-      amnt = 1000;
+      amnt = 100;
+    }
+    if (amnt > 100) {
+      return message.channel.send(message.author + ' Please use a number from 1-100!');
     }
     if (args.length === 0) {
+      message.delete();
       return message.channel.fetchMessages().then(messages => {
-        message.delete();
+        if (amnt > messages.array().size) return message.channel.send(message.author + ' Could not find that many messages!');
         const authMessages = messages.filter(msg => {
           return msg.author.equals(message.author);
         });
@@ -35,30 +39,38 @@ module.exports = {
     }
 
     let canDelOth = message.channel.permissionsFor(message.member).has("MANAGE_MESSAGES");
+    let hasAdmin = message.channel.permissionsFor(message.author).has("ADMINISTRATOR");
     if (!canDelOth) {
       return message.channel.send(message.author + " You don't have permission to use this command!");
     }
 
     if (args[0] === "all") {
-      return message.channel.fetchMessages({
-        limit: amnt
-      }).then(messages => {
-        message.delete();
-        message.channel.bulkDelete(messages).then((deleted) => {
-          message.channel.send('Deleted ' + deleted.size + ' messages from this channel!').then(async sent => {
-            awaitDelete(sent);
+      if (!hasAdmin) return message.channel.send(message.author + "You don't have perission to use this command!");
+      message.delete();
+      try {
+        return message.channel.fetchMessages({
+          limit: amnt
+        }).then(messages => {
+          if (amnt > messages.array().size) return message.channel.send(message.author + ' Could not find that many messages!');
+          message.channel.bulkDelete(messages).then((deleted) => {
+            message.channel.send('Deleted ' + deleted.size + ' messages from this channel!').then(async sent => {
+              awaitDelete(sent);
+            });
           });
         });
-      });
+      } catch (err) {
+        return message.channel.send('There were no messages found in this channel!');
+      }
     }
     const mention = message.mentions.users.first();
     if (!mention) {
       return message.channel.send('That user does not exist!');
     } else {
+      message.delete();
       return message.channel.fetchMessages({
         limit: amnt
       }).then(messages => {
-        message.delete();
+        if (amnt > messages.array().size) return message.channel.send(message.author + ' Could not find that many messages!');
         const mentionMessages = messages.filter(msg => {
           return msg.author.equals(mention);
         });
@@ -73,10 +85,16 @@ module.exports = {
 };
 
 async function awaitDelete(sent) {
-  sent.react("emojiid");
+  sent.react("❎");
+  let emotes = 0;
   sent.awaitReactions(reaction => {
-    if(reaction.emoji.name === "emojiid") {
-      sent.delete(1000);
+    if (reaction.emoji.name === "❎") {
+      emotes++;
+    }
+    switch (emotes) {
+      case 2:
+        sent.delete();
+        break;
     }
   });
 }
